Spell 64-bit constants as the type they are, not as unsigned long long
uint64_t is unsigned long on LP64 and unsigned long long on macOS. A ull
suffix names the latter unconditionally, so every ull literal meeting a
uint64_t operand on Linux was a conversion GCC flags under the
-Wsign-conversion that 7a7193ba turned on with -Werror. The tree therefore
could not build on Linux with tests enabled at all, and the Yocto cross
build only got through because it passes -DBUILD_TESTING=OFF.

The fix is to stop restating the type at the point of use. Where a
declaration already says std::uint64_t the suffix was pure noise and is
deleted; the value is unchanged because a constant that fits is not a
conversion GCC warns about. Verified with a probe against the project's own
warning flags before touching anything.

Where the suffix was load-bearing it becomes the codebase's existing typed
idiom, std::uint64_t{N}, which dbc_parser, road_graph and switchboard
already use. Three sites needed it: ~0ull is int -1 without it, and
2ull * 1024 * 1024 * 1024 overflows int.

That last trap also caught my own first pass. Stripping suffixes turned
10'000ull * 1'000'000 into a literal product evaluated in int, which
overflows to 1410065408 before it ever reaches the uint64_t. The compiler
caught one instance; two neighbours at 200 and 299 million fit today purely
by magnitude, so all three are typed rather than only the one that broke.

Shipping code carried the same defect silently. formatWallClock divides a
std::uint64_t by 1'000'000ull and never warned only because an explicit
static_cast<qint64> sits on the result. The socketcan backend, airplay
timing, two FNV hash loops, an LCG, the bag writer's sizes and the inspect
and bag time formatters were all the same shape. Those are fixed too, since
they would surface the moment someone removed the cast that hides them.

Not touched: literals whose type genuinely is unsigned long long, such as
hex byte patterns and the protowire values paired with LL counterparts.
Normalising those would be churn without a correctness basis.

Verified: the full tree now builds on Linux with tests enabled, which it
could not before. 184 of 185 tests pass. The one failure, msel_test_decode,
reproduces identically at HEAD with these changes stashed, so it is
pre-existing and merely became visible once the suite could run. Every one
of the 76 changed test lines was checked mechanically to differ from its
original by nothing but the removed suffix or a deliberate typed wrapping.

+++ nodes/carplay/simulate.cpp
@@ -216,7 +216,7 @@ bool runSimulation(ZenohBridge& bridge, std::atomic<bool>& stop, int width, int 
                     vf.is_keyframe = isKeyframePacket(pkt);
                     vf.width_px = static_cast<uint16_t>(width);
                     vf.height_px = static_cast<uint16_t>(height);
-                    vf.pts_usec = static_cast<uint64_t>(index) * static_cast<uint64_t>(1'000'000) / static_cast<uint64_t>(fps);
+                    vf.pts_usec = static_cast<uint64_t>(index) * 1'000'000 / static_cast<uint64_t>(fps);
                     vf.data = pkt->data;
                     vf.len = static_cast<size_t>(pkt->size);
                     bridge.publishVideo(vf);
